update new product with working image uploading

+++ db/migrate/20220626225412_create_products.rb
@@ -2,7 +2,7 @@ class CreateProducts < ActiveRecord::Migration[6.1]
   def change
     create_table :products do |t|
       t.string :name
-      t.decimal :price, precision: 8, scale: 2
+      t.integer :price, precision: 8, scale: 2
       t.string :description
       t.string :category
       t.boolean :seasonal
